Added new courses overview page for internal(staff)
